refactor: update user system access through domain model
