+++ src/modules/access-control/data-access/UserSystemAccessRepository.ts
@@ -20,4 +20,19 @@ export default class UserSystemAccessRepository {
       systemRoles: (result.rows[0]?.roles ?? []).map(SystemRole.fromRaw),
     });
   }
+
+  async commit(userAccess: UserSystemAccess): Promise<void> {
+    await this.pool.query(
+      `
+        with delete_step AS (
+            delete from user_system_role
+            where user_id = $1 and role != all($2::system_role[])
+        )
+        insert into user_system_role (user_id, role)
+        select $1, unnest($2::system_role[])
+        on conflict do nothing
+      `,
+      [userAccess.userId, userAccess.systemRoles.map((role) => role.value)],
+    );
+  }
 }
